users: Factor out do_send_password_reset_email.
