+++ zerver/forms.py
@@ -12,16 +12,15 @@ from django.contrib.auth.tokens import PasswordResetTokenGenerator, default_toke
 from django.core.exceptions import ValidationError
 from django.core.validators import validate_email
 from django.http import HttpRequest
-from django.urls import reverse
-from django.utils.http import urlsafe_base64_encode
-from django.utils.translation import get_language, gettext_lazy
 from django.utils.translation import gettext as _
+from django.utils.translation import gettext_lazy
 from markupsafe import Markup
 from two_factor.forms import AuthenticationTokenForm as TwoFactorAuthenticationTokenForm
 from two_factor.utils import totp_digits
 from typing_extensions import override
 
 from zerver.actions.user_settings import do_change_password
+from zerver.actions.users import do_send_password_reset_email
 from zerver.lib.email_validation import (
     email_allowed_for_realm,
     email_reserved_for_system_bots_error,
@@ -31,8 +30,6 @@ from zerver.lib.exceptions import JsonableError, RateLimitedError
 from zerver.lib.i18n import get_language_list
 from zerver.lib.name_restrictions import is_reserved_subdomain
 from zerver.lib.rate_limiter import RateLimitedObject, rate_limit_request_by_ip
-from zerver.lib.send_email import FromAddress, send_email
-from zerver.lib.soft_deactivation import queue_soft_reactivation
 from zerver.lib.subdomains import get_subdomain, is_root_domain_available
 from zerver.lib.users import check_full_name
 from zerver.models import Realm, UserProfile
@@ -357,15 +354,6 @@ class LoggingSetPasswordForm(SetPasswordForm):
         return self.user
 
 
-def generate_password_reset_url(
-    user_profile: UserProfile, token_generator: PasswordResetTokenGenerator
-) -> str:
-    token = token_generator.make_token(user_profile)
-    uid = urlsafe_base64_encode(str(user_profile.id).encode())
-    endpoint = reverse("password_reset_confirm", kwargs=dict(uidb64=uid, token=token))
-    return f"{user_profile.realm.url}{endpoint}"
-
-
 class ZulipPasswordResetForm(PasswordResetForm):
     @override
     def save(
@@ -431,48 +419,9 @@ class ZulipPasswordResetForm(PasswordResetForm):
         except UserProfile.DoesNotExist:
             user = None
 
-        context = {
-            "email": email,
-            "realm_url": realm.url,
-            "realm_name": realm.name,
-        }
-
-        if user is not None and not user.is_active:
-            context["user_deactivated"] = True
-            user = None
-
-        if user is not None:
-            context["active_account_in_realm"] = True
-            context["reset_url"] = generate_password_reset_url(user, token_generator)
-            queue_soft_reactivation(user.id)
-            send_email(
-                "zerver/emails/password_reset",
-                to_user_ids=[user.id],
-                from_name=FromAddress.security_email_from_name(user_profile=user),
-                from_address=FromAddress.tokenized_no_reply_address(),
-                context=context,
-                realm=realm,
-                request=request,
-            )
-        else:
-            context["active_account_in_realm"] = False
-            active_accounts_in_other_realms = UserProfile.objects.filter(
-                delivery_email__iexact=email, is_active=True
-            )
-            if active_accounts_in_other_realms:
-                context["active_accounts_in_other_realms"] = active_accounts_in_other_realms
-            language = get_language()
-
-            send_email(
-                "zerver/emails/password_reset",
-                to_emails=[email],
-                from_name=FromAddress.security_email_from_name(language=language),
-                from_address=FromAddress.tokenized_no_reply_address(),
-                language=language,
-                context=context,
-                realm=realm,
-                request=request,
-            )
+        do_send_password_reset_email(
+            email, realm, user, token_generator=token_generator, request=request
+        )
 
 
 class RateLimitedPasswordResetByEmail(RateLimitedObject):
